backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-allauth==65.0.2
channels==4.1.0
pgvector==0.3.6
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-multipart==0.0.9
openai==1.12.0
python-dotenv==1.0.1
httpx==0.27.2
Pillow==10.4.0
requests==2.32.3
PyJWT==2.9.0
cryptography==44.0.1
